requirements.txt
RPi.GPIO; platform_system == "Linux" and platform_machine == "armv7l"
adafruit-circuitpython-dht; platform_system == "Linux" and platform_machine == "armv7l"
schedule
python-dotenv
pymongo
pydantic
concurrent-log-handler
redis
motor>=3.6
firebase-admin
psycopg2-binary